Removed torch_extension/extension.py file to avoid import issues.
