+++ FSIsolver/extension_operator/extension.py
@@ -328,7 +328,7 @@ import torch
 import torch.nn as nn
 import dolfin as df
 from torch_extension.clement import clement_interpolate
-from torch_extension.extension import poisson_mask_custom, CG1_vector_plus_grad_to_array_w_coords
+from torch_extension.tools import poisson_mask_custom, CG1_vector_plus_grad_to_array_w_coords
 class TorchExtension(ExtensionOperator):
 
     def __init__(self, mesh, model: nn.Module | str, T_switch: float = 0.0, mask_rhs: str | None = None, silent: bool = False):
@@ -429,6 +429,30 @@ class TorchExtension(ExtensionOperator):
             bc.apply(self.u_.vector())
 
         return self.u_
+    
+class TorchExtensionRecord(TorchExtension):
+    def __init__(self, mesh, model, T_switch=0.0, mask_rhs = None, T_record=0.0, run_name="Data0", silent: bool = False):
+        super().__init__(mesh, model, T_switch=T_switch, mask_rhs=mask_rhs, silent=silent)
+
+        # Time to start recording
+        self.T_record = T_record
+
+        # Create time series
+        self.xdmf_input = df.XDMFFile(str(here.parent) + f"/TorchOutput/Extension/{run_name}/harm.xdmf")
+        self.xdmf_output = df.XDMFFile(str(here.parent) + f"/TorchOutput/Extension/{run_name}/torch.xdmf")
+
+        return
+
+    def extend(self, boundary_conditions, params):
+        u_ = super().extend(boundary_conditions, params)
+
+        if params["t"] > self.T_record:
+            self.iter +=1
+            self.xdmf_input.write_checkpoint(self.uh, "input_harmonic_ext", self.iter, df.XDMFFile.Encoding.HDF5, append=True)
+            self.xdmf_output.write_checkpoint(self.u_, "output_pytorch_ext", self.iter, df.XDMFFile.Encoding.HDF5, append=True)
+
+        return u_
+
 
 
 if __name__ == "__main__":
